chore(#4): remove buildPassageTextFromResponse
