--- functions/src/GPromises/api/utils/cleanPassage.ts
@@ -1,7 +1,0 @@
-export const cleanPassage = (passage: string): string => {
-  const cleanedText = passage
-    .replace("¶", "")
-    .trim()
-    .replace(/[^\w?!'"]$/, ".");
-  return /\w$/.test(cleanedText) ? `${cleanedText}.` : cleanedText;
-};
